core: StorageHandlingCommandBase - getEntitiesFromStorageOvfDisk() loop
getEntitiesFromStorageOvfDisk() will enter infinte loop if the
ovfStoreDiskImages list contains only not updated ovf store disks.
This patch changes it to break in that case, avoiding the inifinte loop.
The issue existed since the relevant code was introduced, it was exposed
by https://bugzilla.redhat.com/1412547 which made it  occur frequently.

This patch makes the minimal change to solve the issue as it's going
to be backported to the stable branch, more refactors will be made on
the master branch.

Change-Id: I8669fc6a14879456116fb95ea23737f1ff6672a3
Bug-Url: https://bugzilla.redhat.com/1424755
Signed-off-by: Liron Aravot <laravot@redhat.com>