core: add support for more than one virtio scsi controller
Since every IO Thread has to have it's own virtio scsi controller and the
current code supported only one controller, this patch adds the support for
having more than one.

If the num of io threads of a VM is lowered, the controller Id is re-generated
for this disks.

If the num of io threads is enlarged, the existing disks which already have
addresses are not re-generated, only for the new disks will be assigned to
new controller.

Therefore, in order to enforce the rearrangement of the disks to io threads
which has already been given some address, the disks have to be de-activated
and activated again. The reason is to not to touch the assignement of the disks
to controllers since it influences the disk arrangement inside of the guest.

Change-Id: I5bc8d58237ae28901bf866a50f30e87ee852c980
Bug-Url: https://bugzilla.redhat.com/1422450
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>