core: GlusterServiceSyncJob maps
Anonymously implementing a HashMap just to place values in it is a bad
practice which both "spams" up the classloader with useless anonymous
classes and produces code that's hard to read.

This patch removes the instances of that practice and replaces them with
the better performing Collections.singletonMap where possible or with
an old fashioned HashMap with a series of put calls where it isn't.

Change-Id: Iae6a537076ab389888debe47fc7ed5313fc712b8
Signed-off-by: Allon Mureinik <amureini@redhat.com>