backend: change Discard After Delete logic
Up until now, DAD (Discard After Delete) [1] was not related to Pass
Discard [2], i.e a disk was first discarded before it was deleted only
if its block storage domain's DAD value was true.

That created a situation where a vm disk that had its Pass Discard
enabled was not discarded before it was deleted if its storage domain's
DAD value was disabled. This situation does not make sense because if
the user specifies that he wants "live" discarding (Pass Discard), he
will most likely want to discard the whole disk before it's deleted.

This patch fixes the DAD logic so that a disk will be discarded before
it's deleted if DAD is enabled for its storage domain, or if the same
disk is attached to at least one vm with its Pass Discard enabled.

[1] http://www.ovirt.org/develop/release-management/features/storage/discard-after-delete
[2] http://www.ovirt.org/develop/release-management/features/storage/pass-discard-from-guest-to-underlying-storage/

Change-Id: I06345cd0cebed0ee4ad9ccc930e978807ab47f11
Bug-Url: https://bugzilla.redhat.com/1427543
Signed-off-by: Idan Shaby <ishaby@redhat.com>