webadmin: Fixing unexpected UI behavior in the New External Subnet window
The elements did not fit in on the panel when dns entries were added,
resulting in elements being visible outside the popup window.
Added a scroll panel.
Also enlarged the window, so that the scrollbars fit in without breaking
the panel layout.
The dns size on the subnet panel was made smaller, so that the +/- buttons
fit in. The other use of ExternalSubnetWidget in the AbstractNetworkPopupView
also looks good with the new size.
Screenshots of affected dialogs:
http://imgur.com/a/mGqaM

Change-Id: I66c8fe243299a46b743370627489ae184952af6b
Signed-off-by: Marcin Mirecki <mmirecki@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1425156