core: Monitor devices only for vms that are already in the db
Before this fix vm devices hash was stored in an internal cache
for all vms that were reported by vdsm no matter if they were
present in the engine db at the time.

This caused problems with hosted engine vm that was imported
after the device hash was saved in the cache. Since device monitoring
relies on this cache and there was no change in the hash reported
from vdsm and the hash in the cache the devices weren't imported.

This patch allows monitoring caching info only for vms that already
exist in the db and so solves this issue.

Change-Id: Ib011ac5c7f28de7891b0b6e3c09b107679c1ff29
Bug-Url: https://bugzilla.redhat.com/1425108
Signed-off-by: Jenny Tokar <jtokar@redhat.com>