restapi: Disable filtering by default for non-admin users
The previous patch enabled filtering by default for non-admin users,
like in the master branch. But we don't want to do that in the 4.1
branch, as it is a backwards compatibility breaking change. To avoid
that this patch flips the default value from 'true' to 'false'.

Change-Id: I851d04fa0f1f03b326e40f5434896418d24b7f56
Bug-Url: https://bugzilla.redhat.com/1427987
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 9aeff3498dd09122d50ddcb4db61e6143a03d352)