webadmin: AbstractDiskModel - setDiskProfilesList()
When setDiskProfilesList() is executed, the disk profile set on the disk
is considered when updating the disk profile list.
However, for disk models that doesn't represnt disk which were already 
actually created the profile set in the model isn't the actual disk
profile - therefore if the domain is changed we can update it to the 
newly selected domain profile and not consider it as set profile.

Change-Id: I24d357cdd324817834b8ee2902ec7d3c9add99b6
Bug-Url: https://bugzilla.redhat.com/1360355
Signed-off-by: Liron Aravot <laravot@redhat.com>