core: CloneImageGroupVolumesStructure - determineSourceImageGroup()
determineSourceImageGroup() is used to determine the source image group
(which may differ between the first volume and the rest of the volume
 chain in case the disk is based on a template).
The method had two issues:
1. It returned the image id instead of the disk id.
2. It used image.getImageTemplateId() to determine if the disk is based
on template. When disk snapshots are created their image template id
field is copied from the parent volume, which means that all the chain
has the same value in tht field - only the first volume needs to be
created as based on the template disk. This patch changes the check to
use the template disk id only when the template volume is the parent of
the volume by checking that getParentId() equals to
getImageTemplateId().

Change-Id: Ibf5f6cac89846865bb36e07507390e1a9b21e99a
Bug-Url: https://bugzilla.redhat.com/1429184
Bug-Url: https://bugzilla.redhat.com/1429170
Signed-off-by: Liron Aravot <laravot@redhat.com>