restapi: Fix NPE when requesting ticket on a down VM
When the VM is down, the vm_dynamic-dependent `deriveGraphicsType()` returns
null which caused NPE when forwarding to a subresource of a given console type.

Fixed by forwarding the null graphics type to the backend where the down VM
case is already handled.

In order to do this the ticket invocation code was extracted to the Ticketer
helper class that encapsulates the single operation and is used both by the old
BackendVmResource endpoint and the newer BackendVmGraphicsConsoleResource
subresource endpoints.

Change-Id: I58250c34a5624f8be39b84e7eed1956dc589d995
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1428863
Signed-off-by: Martin Betak <mbetak@redhat.com>