webadmin: change the way of displaying username
This patch changes the way the name of the user is displayed in
webadmin.

Now the username is displayed as 'First Last (username)'. Previously
it was 'First Last (username@authz-name)'. This is valid for permissions
tabs.

For the 'Users' tab we now display in username column just the username
without the authorization provider, because authorization provider
column is just next to username column, so it's redudant information.

Change-Id: I24e6e78b9478e23db088f65d529f5f6e60704bee
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1425323
Signed-off-by: Ondra Machacek <omachace@redhat.com>
(cherry picked from commit d61aaa149ac9e4c2206fb7172bc5eea90ec41e1d)