packaging: services: Limit engine heap size
If ENGINE_HEAP_MIN/ENGINE_HEAP_MAX are too high, set them lower.

They are set by default in new setups to be 25% of available memory,
or 1024MB if machine has < 4096MB, and currently are not changed on
upgrades. If the machine later on has less memory, or if the engine
was restored on a machine with less memory, it might fail. So run it
with lower values and warn in the system log.

Optionally disable this by setting in engine conf:
ENFORCE_ENGINE_HEAP_PARAMS=true

Bug-Url: https://bugzilla.redhat.com/1329119
Change-Id: I2f239691893ad36c723444c956fef0fb755e9c43
Signed-off-by: Yedidyah Bar David <didi@redhat.com>
(cherry picked from commit cd9d7f275ce704ce8436ba213780b7c7ca520d90)