core: Don't wait for jobs to complete during shutdown
When the user requests a shutdown of the engine with the 'systemctl stop
ovirt-engine' command, systemd sends a TERM signal to the service
start/stop script, which in turn sends it to the application server. The
application server then performs its shutdown process. This includes
undeployment of the data sources. When the data source used by the jobs
scheduler is undeployed and there are still jobs pending, they are
retried for ever, and they always fail because there is no longer data
source that they can use. This blocks the application server shutdown
process. The application server is later killed using the KILL signal,
by the start/stop script, after waiting 10 seconds. The net result is
that the engine is stopped anyhow, but systemd reports it as failed, and
some shutdown tasks may have been skipped, because the process was
blocked. To avoid that this patch changes the scheduler implementation
so that it doesn't wait for jobs to finish during shutdown.

Change-Id: I6fc474324d1dc8551933761c3b729d13243b323a
Bug-Url: https://bugzilla.redhat.com/1394617
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>