webadmin: improve Discard After Delete UX
When creating/editing a block storage domain, the user is not warned
when the value of Discard After Delete (DAD) is illegal (the storage
domain contains at least one lun that doesn't support discard).
Since the error comes from the backend, the user doesn't get the chance
to fix the configuration before the popup is closed.

This patch adds UI verifications for DAD to gain a better UX:
1. When enabling DAD, or when selecting a lun while DAD is already
enabled - if the configuration is illegal then a warning message is
shown above the view.
2. When clicking on "ok", DAD is validated so that if the configuration
is illegal then the DAD checkbox is highlighted and the operation is
blocked in the UI level.

Change-Id: I3024fb3ba46ed568cfa0d9d84c15c416da7e335f
Bug-Url: https://bugzilla.redhat.com/1436160
Signed-off-by: Idan Shaby <ishaby@redhat.com>