packaging: Update only invalid values in postgres.conf
Motivation:
The updater of the conf was changing even valid config item. We want to
leave those in place in case someone already took care of that and tweak
the system correctly and not override valid values.

Change:
Make sure valid values are not counted for edition.

Result:
Valid pg values are not amended. If postgres had autovacuum_max_workers
set to 8 and valid value is >= 6 then it is left untouched.

Change-Id: Id06bce50b72544e2f9ee72942a3f6344393f87ce
Bug-Url: https://bugzilla.redhat.com/1388433
Signed-off-by: Roy Golan <rgolan@redhat.com>