restapi: Reduce severity of 'path not found'
Currently when the API doesn't isn't able to calculate the URL path for
a class, it sends an ERROR message to the log, something like this:

  ERROR Can't find relative path for class "o.o.e.a.r.VmDisksResource", will return null

That isn't really an error, rather a hint for developers that something
may be missing in the 'LinkHelper' class. But the ERROR serverity alarms
users. This patch changes the severity to WARNING, which is closer to
the real meaning.

Change-Id: Ib65c904360787d50d22d21fe50747b6ed7453a01
Related-To: https://bugzilla.redhat.com/1410038
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>