core: update cluster fails when VM/Template have wrong values
Failing on the first failed upgrade VM/Template is not informative
for the user.
Instead we don't fail on the first VM/Template failed to upgrade,
we log each VM/Template and the reason in the audit log and preset
all of the failing VMs/Templates for the user to review.

Change-Id: I0968aee9c8388cd3e1352a64035eaae8ba83f28b
https://bugzilla.redhat.com/1418641
Signed-off-by: Shahar Havivi <shaharh@redhat.com>