core: Clarify meaning of memory fields in VDS and VdsStatistics
The memAvailable field returned by VDSM is an approximation of how
much RAM would be available for starting a new VM if all running
VMs used all their allocated memory. But the overhead calculation
is more complicated than that and the engine also has overcommit
rules and so the memAvailable field is not really useful anymore.

Also document the units for related swap fields.

Change-Id: Ib58aecbd9036730f24e38e797f8bc6133c4b7bc7
Signed-off-by: Martin Sivak <msivak@redhat.com>