core: Use a default answer for MockConfigRule
Before this patch, MockConfigRule used the Answers.CALLS_REAL_METHODS
default answer and another Answer to implement the logic of returning
the default config value. This meant that it was still performing a
stubbing operation, which could potentially cause some tests to fail
with an UnnecessaryStubbingException and forcing those tests to use
the Silent runner.

This patch unifies the parts of the stubbing logic to a single
default answer, meaning that this mock now has no stubbing recorded
on it, making it safe to use with the normal MockitoJUnitRunner.

Another side effect this patch has is that it consolidates the
mocking to a single mock object instead of [re]creating it each test,
which should, at least in theory, offer a slight performance
improvement.

Change-Id: If2dcd48aad2596b379eb694d0bf36788ea0f5611
Signed-off-by: Allon Mureinik <amureini@redhat.com>