core: AddClusterCommandTest runner
MockConfigRule used to mock the default values in a way which
couldn't be avoided, making it appealing to use the Silent runner in
this test.

Now that this is not longer the case, this test removes the one
stubbing call that is indeed redundant and restores the stricter
MokcitoJUnitRunner to this test.

Change-Id: I0e718bf186e30a9ebbb2f62350124802936ac50a
Signed-off-by: Allon Mureinik <amureini@redhat.com>