core: AffinityRulesEnforcementManagerTest runner
MockConfigRule used to mock the default values in a way which
couldn't be avoided, making it appealing to use the Silent runner in
this test.

Now that this is not longer the case, this test removes the one
stubbing call that is indeed redundant and restores the stricter
MokcitoJUnitRunner to this test.

Change-Id: I2fbaf531845a26200d5a14436a51afc30715e714
Signed-off-by: Allon Mureinik <amureini@redhat.com>