core: VmValidatorTest runner
MockConfigRule used to mock the default values in a way which
couldn't be avoided, making it appealing to use the Silent runner in
this test.

Now that this is not longer the case, this test removes the one
stubbing call that is indeed redundant and restores the stricter
MokcitoJUnitRunner to this test.

Change-Id: I09f4d29e6b7c20b21e3634c5d826b2656d31c27c
Signed-off-by: Allon Mureinik <amureini@redhat.com>