core: Remove compat from OVF.
Using compat version in the OVF might be misleading since the compat
version of a QCOW volume is determined internally by the
storage domain's version.

Showing a QCOW version in the OVF might mislead
the user to think that previewing this snapshot
will create a volume with this QCOW compat.

Change-Id: I02cf7913653d095012d2f001921dda0e16e8354d
Bug-Url: https://bugzilla.redhat.com/1430447
Bug-Url: https://bugzilla.redhat.com/1432493
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>