core: vm from template - revert flow
When failing on creation of desktop/server vm based on a template
currently there is no rollback of the created disks in case of failure
to complete the entire operation.
For server vms the rollback was never implemented while for desktop vms
the rollback isn't perform since the flow was updated to use the coco
framework (as it was performed using the SPMRevertTask).
As the endWithFailure() method is shared between the flows, this patch
adds the rollback for both.

Note that is this patch is targeted to the 4.1 branch the solution isn't
yet complete - the flow should be rewritten so that the disk is added in
transaction and then the copy flow is initiated.
In case of failure to complete the creation/copy the engine should
attempt
to remove the disk, if the removal fails the disk should remain in
ILLEGAL status so that the user will be able to attempt to delete it to
free the space used by it.

Change-Id: I530f522cd9fe064bb2844ed89f8d4877b25af0f5
Bug-Url: https://bugzilla.redhat.com/1426136
Signed-off-by: Liron Aravot <laravot@redhat.com>