aaa: Add sso client command line registration tool
Add a command line registration tool
ovirt-register-sso-client to register
new sso clients. On running the tool user
is prompted to enter the client id, callback
prefix and the certificate location.

A new entry is created in the sso_clients
table if one does not exist or the existing
one with the same client id is updated.
The client_secret which is written to the
command line should be noted and used by
the client. The client secret in the
sso_clients table is in encrypted format
and is for sso internal use only.

Change-Id: I3b3d5a7c97ff806735f6abc2635cdcf12aadf3f9
Bug-Url: https://bugzilla.redhat.com/1425935
Signed-off-by: Ravi Nori <rnori@redhat.com>