core: fix a regression on import image as template
Recent changes in the add-template flow, to support sealing of the
template, caused a regression in the flow of importing an image from
glance as a template.

Previously, the AddVmTemplate command was not async in this flow (as the
images already exist in the storage domain). Now, we create a callback
anyway which exposes a bug in the detection of whether the VM that the
template is based on exists in the database or not.

This patch provides a fix for this.

Change-Id: I3c61c28be70c1b394b94fac985ce840777035962
Bug-Url: https://bugzilla.redhat.com/1436219
Signed-off-by: Arik Hadas <ahadas@redhat.com>