build: Upgrade checkstyle version
This patch upgrades Checkstyle and Maven Checkstyle Plugin the
project uses to the latests versions (7.6 and 2.17, respectively).

Since the project hasn't upgraded its Checkstyle version for quite
some time (the previously used version was 6.8), this upgrade
introduces some changes that are not backwards compatible and had to
be fixed:
- A suppression was added for all the checks on the generated
  sources, as we can't control their style anyway
- A suppression was added for properties files which weren't analyzed
  with the previous checkstyle version
- ScopeUtils was moved from com.puppycrawl.tools.checkstyle to
  com.puppycrawl.tools.checkstyle.utils
- Its method inInterfaceOrAnnotationBlock was renamed to
  isInInterfaceOrAnnotationBlock

Note that the Checkstyle 7 generation deprecated some classes and
added some new abilities.
Consuming these changes is NOT in the scope of this patch, and will
be handled by the following patches in the series.

Change-Id: I7eae4cbd726e67f3f1b3542208110d831f38a9b7
Bug-Url: https://bugzilla.redhat.com/1433408
Signed-off-by: Allon Mureinik <amureini@redhat.com>