core: getQemuImageInfo should not call teardown on live snapshot.
Once a live snapshot is performed the volume should already be active
in the host and therefore should not be prepared and teardown when
executing getQemuImageInfo.

The fix for this is to check whether the VM, the disks is attached to,
is running, get the host that is running on and call getQemuImageInfo
with that host without the prepare and teardown.

Change-Id: Iadc5d353744cffab8575e5dc4945b707cdca7d59
Bug-Url: https://bugzilla.redhat.com/1433052
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>