restapi: Avoid place holders that aren't replaced
The API documentation uses place holders like {engine-name} and
{product-name} that should be replaced by the corresponding values
during the build process. But this doesn't work correctly because the
HTML generator isn't invoked with the right parameters. That can be
fixed adding the right parameters, or changing the build process so that
it uses the HTML code already included in the artifacts of the model
project. That has the same effect and reduces a bit the time needed for
the build. Note that this mechanism is already used for
'/ovirt-engine/apidoc', this patch just uses the same mechanism for
'/ovirt-engine/api/model', which is now deprecated.

Change-Id: I2f42a0f5d832f7247151cd721dc47c0d2112d3c3
Bug-Url: https://bugzilla.redhat.com/1434056
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>