core: Lock down Extensions' API checkstyle
Commit 879829 upgraded the project's checkstyle version in the root
pom.xml. However, the extensions-api-root module does NOT inherit
from the root module, even though it's in the same project, so it was
not affected by this upgrade.

This patch locks in the checkstyle version of that module to match
the rest of the project.

Change-Id: I329a72215a08ecf20658607a818a2eea2e86ae82
Signed-off-by: Allon Mureinik <amureini@redhat.com>