core: prevent restoring memory from unsupported version
Add a validation to the preview-snapshot flow that prevents the user from
restoring a snapshot with memory that was taken in a version that is no
longer supported.

While it may be debatable, we choose to block the user from trying to
restore the memory of such a snapshot in order to prevent major problems
in the front-end that are caused by that operation.

Change-Id: Ic6a7507e9a266efe8881b2b048795de1c01cf23d
Bug-Url: https://bugzilla.redhat.com/1434941
Signed-off-by: Arik Hadas <ahadas@redhat.com>