core: Allow to deactivate an Inactive domain
The Inactive status is used to mark a domain as in problem. We can conclude
that there's a problem with a domain by inspecting the monitoring results retrieved
from the data center hosts.
Deactivating a domain essentially means to stop monitoring the domain, that's currently
allowed what the domain is Active, but we may want to stop monitor it when there's
a problem with it (and it's in Inactive status) as well.
This patch removes the limitation on doing so and adds attempt to perform ovf update
on the domain OVF stores, as we may still have access to the domain even if it's on Inactive
status (so the OVF update may succeed).

Change-Id: I92bd086020b9f90089b24c67ea1b1c3cf4149d9f
Bug-Url: https://bugzilla.redhat.com/1445753
Signed-off-by: Liron Aravot <laravot@redhat.com>