webadmin: Fix SharedMacPoolComparator
A Comparator's general contract states that:

```
The implementor must ensure that
sgn(compare(x, y)) == -sgn(compare(y, x)) for all x and y.
```

SharedMacPoolComparator breaks this assumption for the case of
comparing two default MacPools. While this may not currently be
possible, it's a bad practice to implement a Comparator in a way that
blatantly breaks Java's contract.

This patch remimplements this Comparator in a way that adheres to
Java's contract and still upholds the intended logic:
1. Default MacPools come first
2. Within each class of "is [not] default", the MacPools
   are sorted lexonumerically by their name.

Change-Id: I9477988dcb680e8d843e8399a4f08a95553e91c1
Signed-off-by: Allon Mureinik <amureini@redhat.com>