webadmin: Fix NetworkInClusterComparator
A Comparator's general contract states that:

```
The implementor must ensure that
sgn(compare(x, y)) == -sgn(compare(y, x)) for all x and y.
```

NetworkInClusterComparator breaks this assumption for the case of
comparing two management networks. While this may not currently be
possible, it's a bad practice to implement a Comparator in a way that
blatantly breaks Java's contract.

This patch remimplements this Comparator in a way that adheres to
Java's contract and still upholds the intended logic:
1. Management Networks(s) come first
2. Within each class of "is [not] management network", the networks
   are sorted lexonumerically by their name.

Change-Id: I4db86c83abdd6ee79bfcf932ab4b1b35a8fae82f
Signed-off-by: Allon Mureinik <amureini@redhat.com>