webadmin: Fix ClusterNetworkModelComparator
A Comparator's general contract states that:

```
The implementor must ensure that
sgn(compare(x, y)) == -sgn(compare(y, x)) for all x and y.
```

ClusterNetworkModelComparator breaks this assumption for the case of
comparing two models where isManagement() returns true. While this
may not currently be possible, it's a bad practice to implement a
Comparator in a way that blatantly breaks Java's contract.

This patch remimplements this Comparator in a way that adheres to
Java's contract and still upholds the intended logic:
1. Management Networks(s) come first
2. Within each class of "is [not] management network", the networks
   are sorted lexonumerically by their name.

Change-Id: Ifedfa7c6cd6b19fbd5126d2174a37740b5738714
Signed-off-by: Allon Mureinik <amureini@redhat.com>