CopyImageGroup/CloneImageGroupVolumesStructure - volume format
CopyImageGroupCommand generates a destination format for the target disk
when performing a disk copy. This format is by the command when copying
a disk with snapshot collapse by passing it to the CopyImage vdsm verb
or when executing the CopyImageGroupWithData command (spdm flow).

The CopyImageGroupWithData command used that format for creating a
collapsed volume, but it passed it to CloneImageGroupVolumesStructure
which is used when cloning the image structure without performing
collapse.

This patch changes CopyImageGroupWithData and 
CloneImageGroupVolumesStructure to not use the generated format when
creating the target image without collapse, but use the original volume
format.

Change-Id: Ie70f847a069a80e3c8e5bc45223da51ef957ed30
Signed-off-by: Liron Aravot <laravot@redhat.com>