engine: Use external providers trust store for neutron API
Use the external providers trust store to connect to external network
providers neutron API. This enables the usage of self signed
certificates by external network providers neutron API, if the
certificate of the external network provider, or its certification
authority, is in engine's external providers trust store.

On the other hand, this change requires that the certificate of the
external network provider, or its certification authority, is added
to the external providers trust store.
The external providers trust store does not contain any default entries,
especially it does not contain the usual root certification authorities.
Via the web user interface of the "Test" functionality the user can
add the certificate of the external network providers neutron API
comfortably.

This patch does not effect the keystone API.

Change-Id: I79a289015329c0fa904b984a8764e73a18da9e5f
Signed-off-by: Dominik Holler <dholler@redhat.com>