restapi: Don't add host statistics by default
In version 3 of the API the hosts used to have the statistics included
inline by default. For example a request like this:

  GET /ovirt-engine/api/hosts/123

Produced a response like this:

  <host id="123" href="/ovirt-engine/api/hosts/123">
    <name>myhost</name>
    ...
    <link rel="statistics" href="/ovirt-engine/api/hosts/123/statistics"/>
    ...
    <statistics>
      <statitstic id="..." href="...">
        <name>memory.total</name>
        <description>Total memory</description>
        ...
      </statistic>
      ...
    </statistics>
 </host>

For the collection of hosts the behaviour was sligthly different: the
statistics were included only if the 'detail' parameter was used to
request it:

  GET /ovirt-engine/api/hosts?detail=statistics

In version 4 of the API the support for this kind of aggregation, and
support for this 'detail' parameter was removed from almost all places,
except from hosts. This has a negative side effect in the SDKs: the XML
tag that contains the statistics doesn't contain a valid 'href'
attribute, and as a result the 'follow_link' method of the SDKs
generates an exception. The workaround for that is to check before
calling the 'follow_link' method. For example, in the Ruby SDK:

  statistics = connection.link?(host.statistics)
    ? connection.follow_link(host.statistics)
    : host.statistics

To avoid this abnormality in the API this patch removes this mechanism
from version 4 of the API, and moves it to the version 3 compatibility
layer.

Change-Id: I5aa6075195e1479746a063acb89f582efa006a14
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>