webadmin: Linq#IPredicate to extend java.util.Predicate
Linq#IPredicate basically serves the same purpose of
java.util.Predicate.
Now that oVirt uses a GWT version that supports JDK 8, IPredicate can
be retrofitted to extend Predicate, where the default implementation
of test would be to just call match.

This is a temporary approach which allows following patches to
incrementally remove usages of IPredicate and replace them with
java.util.Predicate until it can be completely removed.

Change-Id: Idf4b49681ed8bfede233e09d1209d50c48de19d2
Signed-off-by: Allon Mureinik <amureini@redhat.com>