webadmin: Remove Linq#count
Now that Linq#count receives a Collection instead of an Iterable, it
can be replaced with a simple call to Collection#size (or even
Collection#isEmpty when it's just compared to 0).

The worst, this change just cleans up the code. At best, it should
slightly improve performance, as most common collections should have
optimized ways to return their sizes without having to iterate over
the entire collection.

Change-Id: Ib4ab683f62856c8c2969cfcc173a0d75ef3213dc
Signed-off-by: Allon Mureinik <amureini@redhat.com>