webadmin: DataCentersStorageListModel#attachInternal
This patch offers a slight performance improvement in the
attachInternal method.
While the logic of checking if a domain is unattached or present in a
list was not changed, the order of the evaluations was flipped so
that the O(1) is performed before the O(n) operation. If the domain
is unattached, the potentially heavy O(n) check can be short
circuited away.

Change-Id: I15d7e99017b944e322dc7ad6458acf07aa46d514
Signed-off-by: Allon Mureinik <amureini@redhat.com>