webadmin: Reimplement DataCenterGuideModel#updateOptionsNonLocalFS
This patch reimplments DataCenterGuideModel#updateOptionsNonLocalFS
with Java 8 streams instead of the for loop it used to iterate the
hosts.
Besides producing more elegant code, this reimplementation also
improves performance by extracting the cluster IDs to a Set before
attempting to match them to hosts. Thus, for a system with N hosts
and M clusters, this operation is reduced to an O(N) operation
instead of an O(M*N) operation.

The Linq#isClusterItemExistInList method, which is no longer needed
after this rewrite, is thus removed.

Change-Id: If438f9b4327cee8df3596ddb7ac338f4ef9ea25f
Signed-off-by: Allon Mureinik <amureini@redhat.com>