core: ImagesHandler - computeCowImageNeededSize()
When vdsm creates a COW volume with provided initial size the size is
multiplied by 1.1 to prevent a case in which we won't have enough space.
If the source is already COW we don't need the additional space.

determineImageInitialSize() is used to calculate the initial size
and in case that source/dest are both COW divides the space by 1.1 as we
know that the source used space already includes the COW header.

in change I9358478d9f the check for the source image format was
mistakenly removed, leading to that even if the source image format is
RAW we'll divide the initial size by 1.1. As some of the space is used
by the QCOW header we may fail during the copy as we run out of space.

Change-Id: I54923ca1526c39ee793270c28ace6402a558b04a
Bug-Url: https://bugzilla.redhat.com/1438260
Signed-off-by: Liron Aravot <laravot@redhat.com>