webadmin: Inline Linq#findHostByIdFromIdList
Linq#findHostByIdFromIdList is only used in one place,
ModelWithPinnedHost, and as such shouldn't belong in a
general-purpose class like Linq.

Since Java 8's streams provide an elegant, single-statement, way of
achieving the same functionality, this patch removes that method and
inlines its functionality directly where its needed.
As an added bonus, this patch consolidates the filtering and the
sorting of the resulting list, generating a slight performance
improvement.

Change-Id: Ied62d6a2a0a361313e41e5909793ae35a55d1646
Signed-off-by: Allon Mureinik <amureini@redhat.com>