webadmin: Inline Linq#filterProvidersByProvidedType
Linq#filterProvidersByProvidedType is only used in one place,
AsyncDataProvider, and as such shouldn't belong in a general-purpose
class like Linq.

Since Java 8's streams provide an elegant, single-statement, way of
achieving the same functionality, this patch removes that method and
inlines its functionality directly where its needed.
As an added bonus, this patch consolidates the filtering and the
sorting of the resulting list, generating a slight performance
improvement.

Change-Id: I5e7e2bb37529c1d5327ec962e926473acbd2ff4d
Signed-off-by: Allon Mureinik <amureini@redhat.com>