webadmin: Inline Linq#filterDisksByStorageType
Linq#filterDisksByStorageType is only used in one place,
CloneVmModel, and as such shouldn't belong in a general-purpose
class like Linq.

Since Java 8's streams provide an elegant, single-statement, way of
achieving the same functionality, this patch removes that method and
inlines its functionality directly where its needed.
As an added bonus, this patch uses the fail-fast anyMatch method,
which saves the need to iterate over all the VM's disks and construct
a list for them, providing a slight performance improvement.

Change-Id: I634b685fb6be0d3a3afabc025588b1a2488f4e45
Signed-off-by: Allon Mureinik <amureini@redhat.com>