webadmin: Remove Linq#findAllVDSByPmEnabled
Linq#findAllVDSByPmEnabled is only used in one place, HostListModel,
and as such should probably not be part of a general-purpose class
like Linq.

This patch removes that methods and reimplements its logic directly
in HostListModel. As a side bonus, the new implementation uses the
fail-fast allMatch to return false immediately when the first host
that doesn't have power management enabled is found instead of
continuing to iterate over the entire list of hosts.

Change-Id: If6c459331ff21e29f136851cf4f268aad228e4e0
Signed-off-by: Allon Mureinik <amureini@redhat.com>