webadmin: Inline Linq#findAllStorageDomainsBySharedStatus
Linq#findAllStorageDomainsBySharedStatus is only used in one place,
StorageListModel, on a list that is verified to have a single item in
it. As such, it's a poor fit for a general-purpose class like Linq.

This patch removes the aforementioned method and rewrites its logic
in a simple and straight forward way directly in StorageListModel.

Change-Id: I5787aa990e66584d2a125ff7a0c9f40da4a6fcf8
Signed-off-by: Allon Mureinik <amureini@redhat.com>