webadmin: Linq#getStorageDomainsByIds signature
Change Linq#getStorageDomainsByIds to return the interface List
instead of the concrete class ArrayList as per Java's best practices.
Note that this change will require changing several local variables
and/or return values of other methods that rely on it.

This change is done in order to set up its reimplementation in the
following patches.

Change-Id: I2f4be1beaab48a08c7edd5b316c0323d6f3fd203
Signed-off-by: Allon Mureinik <amureini@redhat.com>