webadmin: Linq#filterDisksByType signature
Change Linq#filterDisksByType to return the interface List
instead of the concrete class ArrayList as per Java's best practices.
Note that this change will require changing several local variables
and/or return values of other methods that rely on it.

This change is done in order to set up its reimplementation in the
following patches.

Change-Id: Ice141a7bb6e56e2866784b437a96bc851bb0e714
Signed-off-by: Allon Mureinik <amureini@redhat.com>