webadmin: Linq#filterDisksByType null check
Checking null inside a filter method is unintuitive, and goes against
the convention of similar such methods in the Linq class.
Moreover, all the places the currently call this method already check
for a null input as part of a wider optimization.

This patch removes the null check from the filterDisksByType method
in order to clean up the code and set up further refactoring in the
following patches.

Change-Id: If90304d71315a5806a4b13d0a4e21fc1a7f13e04
Signed-off-by: Allon Mureinik <amureini@redhat.com>