webadmin: Inline Linq#filterNonSnapableDisks
As Linq#filterNonSnapableDisks is only used in one place,
SnapshotMode, it's probably not a good fit for a general purpose
class like Linq, and should be removed from it.
This patch removes this function and inlines the functionality in
SnapshotModel. As an added bonus, this patch consolidates the
filtering and the sorting of the disks list, generating a slight
performance improvement.

Change-Id: Ic052700b9a88fc55e21399241a3d4d1b27e94f9b
Signed-off-by: Allon Mureinik <amureini@redhat.com>