webadmin: Reimplement AbstractDiskModel#updateStorageDomains
This patch reimplements updateStorageDomains so that the list of
storage domains is only traversed once, instead of up to four times
(for regular images - once to get the data domain, once to get the
master domain(s), once to filter only the active domain and then
again to sort the result).

As a result, the methods filterStorageDomainsByStorageDomainType and
filterStorageDomainsByStorageStatus of the Linq class remained
unused, and were subsequently removed.

Change-Id: I42d623c9a6b68db17e3b65c9a2ffc9403c39d426
Signed-off-by: Allon Mureinik <amureini@redhat.com>