core: Properly handle spicevmc redirs on USB support change
After VM run with SPICE + USB enabled there were left several devices of type
'redirdev' which were unmanaged by the engine, and remained in the vm_device
table even after the VM has been reconfigured to VNC (or USB disabled).

This patch modifies the VmDeviceUtils#disableSpiceUsb to also clear the
'redirdev' SPICEVMC devices from the DB. To this end it was necessary to add
new 'REDIRDEV' VmDeviceGeneralType - since the VmDeviceDao always treats the
device 'type' in terms of this enum.

Change-Id: Ib7013e8aa82a6b96802d37d6dcc9babffac37703
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1430688
Signed-off-by: Martin Betak <mbetak@redhat.com>